Move CalcDiffMaskImage to separate file. Remove pixelColorShiftTolerance (not used).
